move all the duplicate post stuff to a new namespace
there's enough stuff here now that it warrants its own separate file.
